Added create residence method in controller. Working on getting the user residence but now getting casting exception on GetUserInfoById that it cannot cast int32 over string...

+++ NETUA2_Egzaminas.DAL/Entities/UserInfo.cs
@@ -15,7 +15,7 @@ namespace NETUA2_Egzaminas.DAL.Entities
         public int Id { get; set; }
         public string Name { get; set; }
         public string Surname { get; set; }
-        public int PersonalID { get; set; }
+        public long PersonalID { get; set; }
         public int PhoneNumber { get; set; }
         public string Email { get; set; }
